Roleplay Information
I was inspired to make this layout after seeing a lot of roleplay information/roleplay comfort memes. I didn't like the layouts of the ones I came across, and felt they were missing somethings (like a writing samples area and a place to put characters someone is most interested in rping at the time)

The code is a bit long, coming in at over 5000k lines, so it might take you awhile to finish. Every section has some prompts in the bullet points to help you know what the section is for if you're unsure or don't know what to list. Remove, add, etc whatever you like though I recommend not doing so, as it could throw off the layout symmetry. The code itself has a color key within it and everything is labeled. Enjoy!